rearranged and modularised to allow easier testing
