+++ geom/CMakeLists.txt
@@ -1,0 +1,7 @@
+# Cmake file for otis
+cmake_minimum_required(VERSION 2.8.4)
+
+# add the geom types
+add_subdirectory(read_cgm)
+add_subdirectory(read_dot)
+
