+++ output/dot/dot_output.hpp
@@ -1,0 +1,8 @@
+#include <vector>
+#include <map>
+
+// print the unique routes
+void print_routes(std::map<int,std::vector<int> > routes);
+
+// convert an int to string
+std::string int_to_string(int integer);
